Specify Job-parameter on CLI and not in config.
In general, it is good design if Values in the config can be set once
and forgot. CLI parameters can change every program invocation though.
So CLI parameters are more fitting for specifying Job-parameters. Not
removing config code because it will most likely be used for other stuff
in the future.