Reimplement search via Twitter's current web UI.
This commit is a first step in rewriting NASTY to no longer use the
legacy UI which we would only recieve through faking an ancient user-agent.
The downside of that approach was that we needed to painstakingly
extract everything from HTML and failed to do so in a few instances,
where we were unable to match the output of the official Twitter API
even for the limited selection of fields that we deemed as relevant to
us.

Using this new approach we only load an HTML stub and than the actual
search results via AJAX requests. THe upside of this is that this
returns the resulting Tweets directly in JSON. In fact in the exact
format that Twitter uses for their developer API. We even receive more
information than if we were using the developer API.

I will remove the old approach and replace it with this new code in the
comming commits.