Remove Pipenv.
Most dependency handling was done in setup.cfg anyways. The last benefit
Pipenv provided was having the Pipfile.lock to specifically denote which
exact versions are used for development. However, this naturaly comes at
the cost of having contributors install pipenv and be familiar with it,
with is rarer than normal virtual environments and incurring a small
performance hit on every execution in the Makefile sicne `pipenv run` is
quite slow.