Rename "executor" command to "batch".
Goal is to be more intuitive.

Previously we had the clash between "submitting requests", "job file",
and "executing jobs/requests", which should be a bit clearer now.

Unfortunately, now we have a clash between "batch" and the "batch-size"
parameter that is forwarded to the Twitter API. Not sure what to do
about this yet.