Add temporary fix to receiving RetryErrors.
These seem to occur once very many Tweets have been received for a
singel request, but to my understanding do not indicate that there are
no more Tweets left for that request. Since I do not know how to
otherwise solve this, I am putting this temporary fix in for now.